+++ lib/src/package_analyzer.dart
@@ -11,7 +11,6 @@ import 'package:path/path.dart' as path;
 
 import 'download_utils.dart';
 import 'internal_model.dart';
-import 'license.dart';
 import 'logging.dart';
 import 'maintenance.dart';
 import 'messages.dart';
@@ -21,7 +20,6 @@ import 'pubspec.dart';
 import 'report/create_report.dart';
 import 'sdk_env.dart';
 import 'tag/tagger.dart';
-import 'third_party/spdx/licenses.dart';
 import 'utils.dart';
 
 class InspectOptions {
@@ -190,21 +188,8 @@ class PackageAnalyzer {
       }
     }
 
-    final licenses = await detectLicenseInDir(pkgDir);
-    if (licenses.isNotEmpty) {
-      tags.addAll(licenses
-          .map((l) => 'license:${l.spdxIdentifier.toLowerCase()}')
-          .toSet());
-      if (licenses.every((l) => fsfLibreLicenses.contains(l.spdxIdentifier))) {
-        tags.add('license:fsf-libre');
-      }
-      if (licenses
-          .every((l) => osiApprovedLicenses.contains(l.spdxIdentifier))) {
-        tags.add('license:osi-approved');
-      }
-    } else {
-      tags.add('license:unknown');
-    }
+    final licenses = await context.licenses;
+    tags.addAll((await context.licenceTags).tags);
 
     List<ProcessedScreenshot>? processedScreenshots = [];
     final screenshotResults = await context.screenshots;